requirements.txt
dwave-ocean-sdk
qubovert==1.2.5
numpy==1.20.3
networkx==2.6.3
matplotlib==3.4.3
PyQt5
scikit-learn==0.24.2